src/requirements.txt
discord.py
tbapy